+++ cmd/codexhub-farfield/process_default.go
@@ -1,0 +1,8 @@
+//go:build !windows
+
+package main
+
+import "os/exec"
+
+func hideConsoleWindow(cmd *exec.Cmd) {
+}